requirements.txt
flask-compress>=1.13
Pillow>=10.0.0
